move `BiDict` to zep dir
